Breadcrumb component
ACCEPTANCE CRITERIA:
As a Content Author: I want to be able to exclude pages from the breadcrumb listing so that structural pages or pages not meant to be navigated to directly are not presented in this context.
As a Site User: I want to be able to see where I am in the site’s content tree so that I understand the context of what I am viewing. I want to be able to navigate to where I have come from in the site so that I can easily jump back and forth between site sections.